user-settable materialization & observation tags (#21452)
## Summary & Motivation

Currently, asset materializations and asset observation events have a
`tags` field. However, users are not able/expected to set values on this
field. This PR changes that.

A notable aspect of what's being proposed here: users can override the
system tags, like those we set to track data provenance at the
materialization level.

The direct catalyst for this PR was a conversation I had with an
ML-focused user, who wanted to be able to customize data provenance
tracking. More info here:
https://github.com/dagster-io/dagster/issues/21436.

I'm not yet convinced this is the right way to handle that need. An
alternative would be something like a `provenance` argument on
`MaterializeResult` and `Output`. But this approach involves less
surface area in the long term.

## How I Tested These Changes