Refactor "invalid asset selection" handling in asset backfills and AMP
Summary:
We were missing a case where a race condition causes the job to be out of sync with user code (a UserCodeProcessError). Also just tweak the logic a bit here so that we actually log the exception when it happens, and more clearly deliniate between the different cases where we might retry.