[DS][1/n] Create SchedulingConditionEvaluationContext (#21499)
## Summary & Motivation

Creates a new SchedulingConditionEvaluationContext. This is essentially a "fresh start" for rebuilding the context object, to make it easier to refactor / delete the current state of the AssetConditionEvaluationContext in the future. It has a greatly reduced set of properties and methods, which can be incrementally built up as needed.

It does keep around a reference to the legacy AssetConditionEvaluationContext, for the sole purpose of letting us keep the existing AutoMaterializeRule implementations untouched. In the eventual future where these are deleted, we can very easily sever the entire legacy context from our new context object.

Note that the weird Union[Scheduling...Context, Asset...Context] thing will go away upstack, so don't worry about that bit.

## How I Tested These Changes