[DS][5/n] Use AssetSlice in AssetConditionResult (#21503)
## Summary & Motivation
Use an AssetSlice object in the internals of this class. Doesn't make a big difference for now, but gives functions that operate over these objects the option to use the AssetSlice API instead of AssetSubset.

Future PRs will change the input type to the create() functions, but that requires larger changes.

## How I Tested These Changes