[DS][23/n] Update SchedulingContext object (#21573)
## Summary & Motivation

This makes a range of changes inspired by comments on https://github.com/dagster-io/dagster/pull/21499.

It hides away any references to AssetConditionX, and simplifies the set of properties available on the context object.

It also adds in a sort of "breadcrumb" thing, where each context object stores a reference to its parent context object. No existing rules currently take advantage of this, but the soon-to-be-created "NewerThanChild" condition (name pending) will be a rule that gets applied against dependencies, but needs to know which child it is comparing itself to. i.e. you would have something like:

SchedulingCondition.any_dep_matches(SchedulingCondition.newer_than_child())

## How I Tested These Changes