[DS][24/n] Remove all AssetSubset references from the context and result objects (#21612)
## Summary & Motivation

As title -- for the new codepaths, helps avoid a lot of convert_to_valid_asset_subset() business.

I didn't bother updating the AutoMaterializeRule side of the world in a nicer way, as this code will eventually be removed anyway.

Had to add some new methods to replace the & / | / - operators, this time properly labeled with "compute_"

## How I Tested These Changes