[ui] Metadata entry plots for asset checks (#21647)
## Summary & Motivation

This PR adds a "Plots" section to the asset checks page, displaying
numeric metadata found in asset check evaluations. The plots are always
on a time axis (for now), but re-use the asset materialization code
which should make it easy to show them by partition later.

In a follow-up, I'll rename the graph components from
`<AssetMaterializationGraphs />` to `<MetadataEntryPlots />` to reflect
this checks use and the "plots" naming - just didn't want to clutter
this one.
 
Test asset:

```
def asset_with_check_in_same_op():
    yield Output(1, metadata={"percent_complete": random.random(), "a_value": 3})
    random.seed(time.time())
    yield AssetCheckResult(check_name="random_fail_check", metadata={"percent_complete": random.random(), "a_value": 5}, passed=random.choice([False, True]))

```

Screenshots!


![image](https://github.com/dagster-io/dagster/assets/1037212/3e317fdc-7f5c-41a3-abc2-a005eca15438)


Empty state:

![image](https://github.com/dagster-io/dagster/assets/1037212/8ed75c9e-35da-40a2-a4bb-116ddf02f166)


## How I Tested These Changes

This PR makes several changes to the padding and borders on asset plots
(we removed the borders), so I checked these screens below to make sure
they still look good.



![image](https://github.com/dagster-io/dagster/assets/1037212/8f8e7cdc-9f91-4526-84b2-b5d96ef9587b)

![image](https://github.com/dagster-io/dagster/assets/1037212/78914a92-50b3-4cdb-91f7-8b2bb55f0517)

![image](https://github.com/dagster-io/dagster/assets/1037212/83bcbae1-fe6f-4ded-9182-28bd006d55b4)


This PR also fixes the timestamps in the X axis so that they reflect the
user's timezone preference and the rendering we use everywhere else. I
tested this by opening the settings modal and making sure that fiddling
with the settings adjusts the presentation:


![image](https://github.com/dagster-io/dagster/assets/1037212/e5bcb140-3af7-4d80-8693-fc9a1ed6c872)

---------

Co-authored-by: bengotow <bgotow@elementl.com>