Allow metadata sets to have fields that are subtypes of `MetadataValue` (#21271)
## Summary & Motivation

The current implementation of `NamespacedMetadataSet.extract` assumes
that the types of all fields are not subclasses `MetadataValue`. When it
comes across a value of type `MetadataValue`, it auto-extracts the
primitive value from inside it.

This is problematic if you want to specify that a field on a metadata
set is, e.g., a URL. The return type of `UrlMetadataValue.value` is
`str`.

This PR makes the implementation of `NamespacedMetadataSet.extract` more
sophisticated: it now looks at the type annotation on the field to
determine whether to pull out the value.

Alternatives to consider here:
- Require that the `value` property on all subclasses of `MetadataValue`
returns a unique type. E.g. make `UrlMetadataValue` return a `Url`
instead of a `str`. This would reduce the need for
`NamespacedMetadataSet` subclasses to have values with types that are
subclasses of `MetadataValue`. However, this would be a breaking change,
as well as require introducing new types like `Url`.
- Require that all fields on metadata sets have types that are
subclasses of `MetadataValue`. However, barring primitive types would
make the API significantly more clunky to use.

## How I Tested These Changes