feat(looker): represent looker views as assets (#21687)
## Summary & Motivation
Model the basic Looker view, where a view is queried off of an existing sql table.

In this case, we can determine the Looker view's upstream dependency using [`sql_table_name`](https://cloud.google.com/looker/docs/reference/param-view-sql-table-name). In the case that `sql_table_name` is not defined, then use the name of the Looker view as the upstream dependency.

In a downstream PR, we'll determine the dependency of Looker views that rely on sql specified in [`derived_table`](https://cloud.google.com/looker/docs/reference/param-view-derived-table).

## How I Tested These Changes
pytest