in upath_io_manager, set no path metadata when obj is None (#21337)
## Summary & Motivation

Usually, when an op or asset returns `None`, it means that the user
didn't intend to use an IO manager, but ended up with one because they
didn't set `-> None` as their return type annotation. In these cases,
seeing a "path" metadata value can be very confusing, because usually
the body of their op is actually writing data to a different location.

We've discussed solutions at length that involved bypassing the IO
manager, but ultimately backed off from them due to backwards
compatibility concerns.

This PR proposes a smaller step: omitting the metadata when the
outputted in object is `None`. This avoids the confusion in the common
case. There's a tradeoff here: this comes at the cost of consistency.
And it means that, in the rare cases where people actually intend to
return `None` and have it written to a file, they won't see the metadata
for that file.

## How I Tested These Changes