Standardize existing row count metadata to use `dagster/row_count` key (#21524)
## Summary

Moves the existing row count metadata to be a subset of the `TableMetadataSet`, so it is prefixed as `dagster/row_count`. This identifies it as a special metadata type, lets us specially handle it in the UI, and also ensures it remains the same across integrations.

Introduces `dagster/partition_row_count` for partitioned asset materializations, as well.

## Test Plan

Brief unit test of `TableMetadataSet`, update other unit tests.