Add TreatAsResourceParam; Use it for K8sPipeClient (#21719)
## Summary & Motivation

When developing Pipes, we developed a pattern where we would implement classes prefixed with an underscore, with the actual name reserved for a `ResourceParam`-annotated symbol of that class.

This eliminated the need for `ResourceParam` usage when using Pipes client.

While convenient, this broke docs which is unfortunate, and also degrades the in-IDE experience.

The proposal here is to add a new marker interface `TreatAsResourceParam`.

Now you can write the following:

```python
class PipesK8sClient(PipesClient, TreatAsResourceParam):
   ...
```

And use the type directly as an argument.

## How I Tested These Changes

![Screenshot 2024-05-08 at 6.54.54 AM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/wS38mdcdU6aAeJobBdry/15573bda-bcf5-431c-bf6e-c92619409304.png)

BK