doc: serverless to hybrid transition (#21604)
## Summary & Motivation

This change improves the instructions for deploying Dagster+ in serverless environments. It includes additional steps and details you need to consider when switching from Serverless to Hybrid. This includes updating the code location configurations and images, modifying the build process, and replacing Serverless deployment features with their Hybrid equivalents. This change assists users in adapting their deployment workflows in the most beneficial way.

## How I Tested These Changes

- CI/CD + doc preview