feat: pass more options to dagster-aws Resources (#21637)
## Summary & Motivation

I am using dagster with localstack for dev testing, and I am unable to
pass `endpoint_url` variable to the `SSMResource` as with the
`S3Resource`. This PR will allow for more configuration options to be
passed to resources that support it.

- `CloudwatchLogsHandler`
- `ECRPublicClient`
- `SecretsManagerResource`
- `SSMResource`


## How I Tested These Changes

The unit tests for the library passed. As the resources that are being
changed, all call `boto3.session.Session.client()` verified that the
options are in the [documentation for
boto3](https://boto3.amazonaws.com/v1/documentation/api/latest/reference/core/session.html#boto3.session.Session.client).

---------

Co-authored-by: Mathieu Larose <mathieu.larose@gmail.com>