Minimize workspace snapshot dependency from GrapheneAssetLatestInfo
Summary:
The medium term plan here is to have GrapheneAssetNode be the field that requires definition-level data, and the imperfectly named GrapheneAssetLatestInfo to handle all the fields that require access to the

In the short-term, we still need the nodes in the case where the most recent run for that asset is in progress. But that should still be a visible improvement over fetching the nodes for every asset on every query.