[docs] - Add examples page for Schedules (DOC-191) (#21735)
## Summary & Motivation

This PR experiments with adding an **Examples** page for Schedules. It
pulls some examples from the current Schedules concept page and
simplifies the explanations to make it easier to skim. The focus of this
reference is not to teach users how to do X, but to quickly demonstrate,
provide relevant links to jump to, and let them get back to work.

I've added a summary table to each example that contains:

- Any notes, tips, etc. about the example
- Links to related documentation. For example, where the code is used
and has a more learning-focused explanation
- A list of APIs used in each example

![Screenshot 2024-05-08 at 5 27 36
PM](https://github.com/dagster-io/dagster/assets/16615212/a9c04055-5a0b-4a3d-b294-b011c7fa337b)

## How I Tested These Changes

👀