feat(looker): represent looker views dependencies from derived tables (#21689)
## Summary & Motivation

Introduce `sqlglot` to parse Looker views that are defined with a
`derived_table`.

The annoying part here is that the SQL here is potentially templated:
https://cloud.google.com/looker/docs/sql-and-referring-to-lookml#substitution_operator_.
In this case, don't give it upstream dependencies.

In a future PR, we can potentially look at a way to use regex to
template the names into file, using a context dictionary assembled from
the total aggregate of `*.view.lkml` files.

## How I Tested These Changes
pytest

- SQL with the `$` operator will have an empty set of upstream
dependencies
- SQL without the `$` operator but cannot be parsed by `sqlglot` will
have an empty set of upstream dependencies