Fix the strftime tz argument and make optional (#21475)
## Summary & Motivation

The check in this method if the timezone name is the same on the datetime as what was passed in was incorrect. It was always false. It's hard to write correctly though honestly. There isn't a method on tzinfo or datetime that exposes the full timezone name that was used to create the `tzinfo`. 

Ideally, strftime shouldn't really be taking a datetime and a tz. Callers should just convert the datetime to the tz before calling the method. Instead of doing that larger refactor, I'm slipping in an optimization: Many of the methods on TimeWindowPartitionDef call this strftime method with timestamps that were generated by the cron iterator. All datetimes from that iterator are already using `self.timezone` for the `tzinfo`. 

## How I Tested These Changes

bk