Fix return type of load_assets_from_modules
Concrete functions should specify the actual type
of what they return. In this case, these functions
all returns lists, NOT a generic Sequence.

In general, functions and classes in a library should
be as generic as possible in what they accept as input
but be as specific as possible in what they return
as output.

These returns lists. If I try to append or manipulate
the returns, they have the incorrect type of Sequence.
A sequence does not have .append for instance. It
cannot be concatenated with a list of AssetDefinitions,
etc. If using strict type checking, which we do, I
cannot even easily combine the output of these functions
with assets imported "manually". I would get the error
 list[AssetsDefinition] is incompatiable with Sequence[...].

It is quite annoying to have to `typing.cast` the output
of library code from Dagster into its actual type just to
get my own type checking to work. It is made harder by the
fact that CacheableAssetsDefintion is not exported by default,
so I have to import it, and cast the output to
`list[AssetsDefinition | SourceAsset | Cacheable...]]`.

This fixes the return of these functions to the actual type,
meaning that they are still `Sequence[Union[A, B, C]`, and will
naturally pass the type check for any functions that works on a
generic Sequence[Union[A, B ,C]`, but
they are now also more concretely `List[Union[A, B, C]` so
one can actually use the returned objects in their own user code.

Change-Id: I13a06494c69c622dd65b4349d9538f94f1b3fc1a