Fix disambiguation routes (#21811)
## Summary & Motivation

In a previous PR I removed a waterfall on the WorkspaceQuery since many
of the child routes of `WorkspaceRoot` do not actually depend on it.

Well in removing that I made a bug that was previously pretty
intermittent / infrequent happen all the time. In particular there were
routes that were depending on the permissions query without actually
waiting for it. Instead those routes happened to be waiting on the
WorkspaceQuery which would normally come back after the permissions
query hence why the routes worked before.

The reason this bug happened intermittently before is because sometimes
the WorkspaceQuery would come back before the permissions query causing
this bug.

To fix this I make the routes wait for permissions to load directly.


## How I Tested These Changes

Loaded this page Via app-cloud-proxy and saw that we went to the launch
pad:
http://localhost:3001/elementl/prod/locations/dagster_open_platform/pipeline_or_job/dagster_quickstart_validation_job/playground/setup?config=%7B%7D%0A&tags%5B0%5D%5Bkey%5D=dagster%2Fsensor_name&tags%5B0%5D%5Bvalue%5D=dagster_quickstart_validation_sensor&tags%5B0%5D%5B__typename%5D=PipelineTag