only fetch observations in get_latest_data_version_record if required (#21809)
## Summary & Motivation
The current implementation of `get_latest_data_version_record` would
always fetch the latest observation AND the latest materialization. It
would then return one or the other, preferring the materialization in
the ambiguous case.

This PR short-circuits the ambiguous case such that we only fetch the
observation when the materialization is not present. This effectively
halves (from 2 calls => 1 call) the round-trips to storage in the most
common case (materializable assets).

Additionally, this PR replaces calls to the generic `get_event_records`
API with the narrower `fetch_X` APIs.

This PR is orthogonal to
https://github.com/dagster-io/dagster/pull/21798, which replaces calls
to `get_latest_data_version_record` with calls to the batchable
`get_asset_records` in the simple case where storage_id filters and
partition filters are not applied.

Should note: this does not change the logic of which record is returned.
The API name is a little misleading because we will return the latest
materialization record even if the observation record is more recent for
materializable assets with both types of records.

## How I Tested These Changes
BK