[bug] Fix issue causing ReconstructableJobs containing certain types of metadata to error (#21819)
## Summary & Motivation

Resolves: https://github.com/dagster-io/dagster/issues/21815

Note that while the test involved here is referencing
`CacheableAssetsDefinition`, this same error would occur with regular
assets. Any asset using a `MetadataValue` which contains within it a
mutable object (which would be table schema and table column lineage
which contain lists I believe) would encounter a similar error.

After switching the core `MetadataValue` class from a `NamedTuple` to a
pydantic model, the code we use to generate hashes for our serializable
objects stopped handling them. This PR fixes this.

## How I Tested These Changes

After updating the test to include metadata of this type, observed the
mentioned error. These changes resolved it.