[DS][31/n] Create static constructor for eager condition (#21737)
## Summary & Motivation

This creates a static constructor for an analog to the current-day eager condition. It is not identical, as it does not consider historical time partitions. It also manages state differently.

In the current world, the "parent_newer" analog actually encodes a lot of different information. From its perspective, an asset partition is True if:

A parent has been materialized more recently than the asset, and the run that materialized it does not target this asset,
AND this asset hasn't been requested for materialization by AMP since that run came in

This is a pretty confusing thing to wrap your head around, but it does have two good qualities:

it avoids repeatedly requesting the downstream asset for materialization while the asset computes (techincally, the parent will remain newer than the child until the child successfully materializes, which may in practice span a large number of scheduling evaluations)
it avoids repeatedly materializing the asset in the case that the requested materialization failed (similar to the above issue)
This policy handles issue (1) by explicitly avoiding materializing this asset if a run is currently in progress, but does not handle case (2).

This will need to be handled with a recently_requested condition, which can keep track of the asset partitions which have been requested within some timedelta.



## How I Tested These Changes