[DS][32/n] Update UpdatedSinceCronCondition (#21677)
## Summary & Motivation

First, realized that this wasn't actually tested. Added real tests, and updated the implementation to be slightly more efficient. In particular, previously whenever any partition of a given asset was updated since the previous tick (not near a cron boundary), we'd end up doing a somewhat-inefficient db query to get all the materialization events since the previous cron tick. Instead, we can just use the (probably already-fetched) information about materializations since the previous tick and keep track of that. Not a game changer or anything, but a bit nicer.

## How I Tested These Changes