feat(dbt): omit `fqn:` when constructing a subssetted execution (#21851)
## Summary & Motivation
Save 4 characters per each selected dbt asset and check.

Before, we were being explicit that `fqn:` was being used. Instead, we
can just let dbt figure that out. From
https://docs.getdbt.com/reference/node-selection/methods, `fqn:` is one
of the default selectors anyways.

This isn't going to make a huge dent in the most diabolical of cases,
but might as well cash in on some character savings.

## How I Tested These Changes
pytest