Hoist the asset daemon context out of the LegacyRuleEvaluationContext (#21893)
## Summary & Motivation

This goal of this is to introduce a user-facing testing API for scheduling conditions that can be executed outside of the Asset Daemon. The process here is to extract the asset condition evaluation logic out of the `AssetDaemonContext` and into its own class that does not reference `AssetDaemonContext`. The general process is to hoist the properties on the `AssetDaemonContext` that these code paths need out of that class, and then change code to reference the constituent properties instead of the context.

This PR eliminates the `AssetDaemonContext` property from `LegacyRuleEvaluationContext`

## How I Tested These Changes

BK