Add quickstart example for dbt & Dagster (#21240)
## Summary & Motivation

This PR adds a quickstart example for dbt & Dagster. A user that wants
to skip the tutorial can copy and paste the code included in the example
and run `dagster dev`.

This example includes the new experimental class `DbtProject`. Another
PR will add more DbtProject examples in the dbt docs and notes that the
feature is still experimental.

## How I Tested These Changes

BK 
+
local
```
make mdx-full-format
make next-watch-build
```