[dagster-dbt] add translator option to attach model files as code reference metadata (#21888)
## Summary

Adds the option to a custom `DagsterDbtTranslator` to attach code
reference metadata pointing to the dbt model files for each asset.

Unfortunately, this requires that we also get the `project_dir ` from
the user, since that is only available at run time as part of the
resource. Maybe there's a good way around this? The manifest used to
have a `root_path` field, but this [was recently
removed](https://github.com/dbt-labs/dbt-core/issues/6171).
 
```python
@dbt_assets(
    manifest=...,
    dagster_dbt_translator=DagsterDbtTranslator(
        settings=DagsterDbtTranslatorSettings(attach_sql_model_code_reference=True)
    ),
    project=DbtProject(...)
)
def my_dbt_assets():
  ...
```

## Test Plan

Unit tests.