[3/5] [Run Timeline] Add hourly data cache (#21934)
## Summary & Motivation

Add an hourly data cache to the Run Timeline. 
We only cache completed run data and we only request data for time
ranges that we don't already have data for.


- how do we handle settling ongoing runs in to the completed bucket?
- - when an ongoing run goes into completed it should have an
updatedTimestamp that belongs to a range we havent fetched yet

Updated the completed run bucketing query logic to fetch updatedAfter ->
updatedBefore timestamps to avoid the overfetching that the current
updatedAfter -> createdBefore buckets make.


![image](https://github.com/dagster-io/dagster/assets/2286579/af0b2c02-f189-4332-95a2-5cacba95a6d3)


updatedAfter -> updatedBefore buckets only capture range 3 and 4 boxes
though:

![image](https://github.com/dagster-io/dagster/assets/2286579/c566547a-43b9-4d83-bd07-5e5b9e1285be)


To capture range 2 and range 1 we need runs that were createdBefore the
right boundary and updated
after the right boundary. Instead of querying for these runs
specifically we rely on the fact that we have already queried for all
the future buckets since we always paginate backwards.

So we subscribe to the cache for that data and then filter for runs that
started before the right boundary of the time range


## How I Tested These Changes
Jest tests.

Left the page open for half an hour and saw in progress runs transition
to completed.

Switched between 1hr / 6hr / 12hr / 24hr views and saw we only request
time ranges that are missing.